Moved addComment and refactored
