+++ frontend/src/core/comments/index.js
@@ -1,5 +1,7 @@
 /* @flow */
 
+export { default as actions } from './actions';
+
 export { default as AddComment } from './components/AddComment';
 export { default as Comment } from './components/Comment';
 export { default as CommentsList } from './components/CommentsList';
