+++ frontend/src/modules/teamcomments/actions.js
@@ -1,10 +1,6 @@
 /* @flow */
 
-import NProgress from 'nprogress';
-
 import api from 'core/api';
-import * as history from 'modules/history';
-import * as notification from 'core/notification';
 
 import type { TeamComment } from 'core/api';
 
@@ -55,34 +51,8 @@ export function get(entity: number, locale: string): Function {
 }
 
 
-export function addComment(
-    entity: number,
-    locale: string,
-    pluralForm: number,
-    translation?: number,
-    comment: string,
-): Function {
-    return async dispatch => {
-        NProgress.start();
-
-        await api.comment.add(entity, locale, comment, translation);
-
-        dispatch(notification.actions.add(notification.messages.COMMENT_ADDED));
-        if (translation) {
-            dispatch(history.actions.get(entity, locale, pluralForm));
-        }
-        else {
-            dispatch(get(entity, locale));
-        }
-
-        NProgress.done();
-    }
-}
-
-
 export default {
     get,
     receive,
     request,
-    addComment,
 };
